pyproject.toml
[project]
name = "StoryTime"
version = "0.0.1"
description = "A TUI app that generates short stories and AI images from prompts using a language model."
authors = [
    { name = "wom" }
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "google-genai>=1.23.0",
    "pillow>=11.2.1",
    "textual>=0.54.1"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "ruff>=0.12.1",
    "pip>=25.1.1"
]

[project.scripts]
storytime = "storytime.StoryTime:main"
